Flushing out front end for V1
got a little carried away this weekend and added plenty of non-V1 material.
api games is tested and is good
api pieces is currently untested but, based of tested curls

components are structured as
app - gameHall - CreateGame & IndexGame
|| parallel / siblings
app - GameLobby - GameBoard

the IndexGame component is a V2.5+ feature and is currently not fully functional (using href instead of Links, known issue)
to be taken out of paths for V1.

the same is true for the GameLobby component

so will be
app - gameHall - CreateGame where gameHall renders CreateGame as a child
|| parallel / siblings
app - GameBoard

GameBoard component is in development and is THE V1. MVP
Currently adding the state with set functions, an on-mount to initialize the game data, and an update piece function to do a patch AND an index request to return the new piece array instead of splicing the piece state array after a patch and hoping it works.