flushing out the game board mvp
hoisted the game state into the app with methods passed in props.
the GameBoard is a child of Game GameInstance

component structure:
app - game hall - create game & index game
|| parallel
app - game instance - game board

gameBoard can be given a new arg for size to draw the board. Currently 7 is hardcoded in to the loops.