Add Some Piece integration
added new component, GameCell, that has replaced the board cell on the gameboard component. It has 3 'layers' as state variables, 1 for the x and y coordinate object, 1 for the piece data if there is one and one for others, currently 'building layer': unsure of how I want to use this if at all in the future.

GameCell displays the pieces name if there is one on the cell.
the board now updates when a new piece is added, no refresh necessary HOWEVER ! this was accomplished with the UNSAFE_componentWillReceiveProps. A refactor will have to take place to remove the need for this in the gameBoard component as development continues.

Also added to the state on game board is a selected array, hoping to be able to use a state function to change the display colors of the children components, from other children. ( I can move this piece to these squares, or attack that other piece etc...)